renames moves/index to moves/moves
